tests/pyproject.toml
[project]
name = "k8s-node-perf-evaluator-tests"
version = "0.1.0"
description = "Tests for K8s Node Perf Evaluator"
authors = [
    {name = "Nadun De Silva",email = "nadunrds@gmail.com"}
]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest (>=8.3.5,<9.0.0)",
    "docker (>=7.1.0,<8.0.0)",
    "mypy (>=1.16.0,<2.0.0)",
    "types-docker (>=7.1.0.20250523,<8.0.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "types-pyyaml (>=6.0.12.20250516,<7.0.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.12"
